+++ include/nstd/gl/texture.h
@@ -1,6 +1,6 @@
 #ifndef NSTD_GL_TEXTURE_H_INCLUDED
 #define NSTD_GL_TEXTURE_H_INCLUDED
 #include "../nstd.h"
-#include "texture/texture.h"
+#include "texture/format.h"
 #include "texture/view.h"
 #endif